Restructuring code.
