+++ write.go
@@ -1,0 +1,83 @@
+// Package scribble is a tiny JSON database
+package scribble
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+)
+
+// Write locks the database and attempts to write the record to the database under
+// the [collection] specified with the [resource] name given
+func (d *Driver) Write(collection, resource string, v interface{}) error {
+	// ensure there is a place to save record
+	if collection == "" {
+		return fmt.Errorf("Missing collection - no place to save record!")
+	}
+
+	// ensure there is a resource (name) to save record as
+	if resource == "" {
+		return fmt.Errorf("Missing resource - unable to save record (no name)!")
+	}
+
+	// create mutex on collection
+	mutex := d.getOrCreateMutex(collection)
+	mutex.Lock()
+	defer mutex.Unlock()
+
+	// create full paths to collection, final resource file, and temp file
+	dir := filepath.Join(d.dir, collection)
+	fnlPath := filepath.Join(dir, resource+".json")
+	tmpPath := fnlPath + ".tmp"
+
+	// create collection directory
+	if err := os.MkdirAll(dir, 0755); err != nil {
+		return err
+	}
+
+	// marshal input
+	b, err := json.MarshalIndent(v, "", "\t")
+	if err != nil {
+		return err
+	}
+
+	// write marshaled data to the temp file
+	if err := ioutil.WriteFile(tmpPath, b, 0644); err != nil {
+		return err
+	}
+
+	// move final file into place
+	return os.Rename(tmpPath, fnlPath)
+}
+
+// Delete locks that database and then attempts to remove the collection/resource
+// specified by [path]
+func (d *Driver) Delete(collection, resource string) error {
+	// create full path to resource
+	path := filepath.Join(collection, resource)
+
+	// create mutex
+	mutex := d.getOrCreateMutex(collection)
+	mutex.Lock()
+	defer mutex.Unlock()
+
+	// create full path to directory
+	dir := filepath.Join(d.dir, path)
+	switch fi, err := stat(dir); {
+	// if fi is nil or error is not nil return
+	case fi == nil, err != nil:
+		return fmt.Errorf("Unable to find file or directory named %v\n", path)
+
+	// remove directory and all contents
+	case fi.Mode().IsDir():
+		return os.RemoveAll(dir)
+
+	// remove file
+	case fi.Mode().IsRegular():
+		return os.RemoveAll(dir + ".json")
+	}
+
+	return nil
+}
